+++ core/src/alu/mod.rs
@@ -29,7 +29,7 @@ pub struct AluEvent {
     pub shard: u32,
 
     /// The channel number, used for byte lookup table.
-    pub channel: u32,
+    pub channel: u8,
 
     /// The clock cycle that the operation occurs on.
     pub clk: u32,
@@ -51,7 +51,7 @@ pub struct AluEvent {
 
 impl AluEvent {
     /// Creates a new `AluEvent`.
-    pub fn new(shard: u32, channel: u32, clk: u32, opcode: Opcode, a: u32, b: u32, c: u32) -> Self {
+    pub fn new(shard: u32, channel: u8, clk: u32, opcode: Opcode, a: u32, b: u32, c: u32) -> Self {
         Self {
             lookup_id: 0,
             shard,