+++ core/src/utils/mod.rs
@@ -1,4 +1,5 @@
 mod buffer;
+pub mod concurrency;
 mod config;
 pub mod ec;
 mod logger;