+++ recursion/circuit/src/constraints.rs
@@ -167,11 +167,10 @@ mod tests {
     use p3_baby_bear::DiffusionMatrixBabyBear;
     use p3_challenger::{CanObserve, FieldChallenger};
     use p3_commit::{Pcs, PolynomialSpace};
-    use serde::{de::DeserializeOwned, Serialize};
     use sp1_core::{
         stark::{
-            Chip, Com, DefaultProver, Dom, MachineProver, OpeningProof, PcsProverData,
-            ShardCommitment, ShardMainData, ShardProof, StarkGenericConfig, StarkMachine,
+            Chip, Com, CpuProver, Dom, MachineProver, OpeningProof, PcsProverData, ShardCommitment,
+            ShardProof, StarkGenericConfig, StarkMachine,
         },
         utils::SP1CoreOpts,
     };
@@ -208,7 +207,6 @@ mod tests {
         OpeningProof<SC>: Send + Sync,
         Com<SC>: Send + Sync,
         PcsProverData<SC>: Send + Sync,
-        ShardMainData<SC>: Serialize + DeserializeOwned,
         SC::Val: p3_field::PrimeField32,
         <SC as sp1_core::stark::StarkGenericConfig>::Val:
             p3_field::extension::BinomiallyExtendable<4>,
@@ -296,7 +294,7 @@ mod tests {
         let mut runtime = Runtime::<F, EF, DiffusionMatrixBabyBear>::new_no_perm(&program);
         runtime.run().unwrap();
         let machine = A::machine(config);
-        let prover = DefaultProver::new(machine);
+        let prover = CpuProver::new(machine);
         let (pk, vk) = prover.setup(&program);
         let mut challenger = prover.config().challenger();
         let proof = prover