+++ zkvm/entrypoint/src/syscalls/ed25519.rs
@@ -4,9 +4,14 @@ use core::arch::asm;
 /// Adds two Edwards points.
 ///
 /// The result is stored in the first point.
+///
+/// ### Safety
+///
+/// The caller must ensure that `p` and `q` are valid pointers to data that is aligned along a four
+/// byte boundary.
 #[allow(unused_variables)]
 #[no_mangle]
-pub extern "C" fn syscall_ed_add(p: *mut u32, q: *mut u32) {
+pub extern "C" fn syscall_ed_add(p: *mut [u32; 16], q: *const [u32; 16]) {
     #[cfg(target_os = "zkvm")]
     unsafe {
         asm!(
@@ -26,6 +31,11 @@ pub extern "C" fn syscall_ed_add(p: *mut u32, q: *mut u32) {
 /// The second half of the input array should contain the compressed Y point with the final bit as
 /// the sign bit. The first half of the input array will be overwritten with the decompressed point,
 /// and the sign bit will be removed.
+///
+/// ### Safety
+///
+/// The caller must ensure that `point` is valid pointer to data that is aligned along a four byte
+/// boundary.
 #[allow(unused_variables)]
 #[no_mangle]
 pub extern "C" fn syscall_ed_decompress(point: &mut [u8; 64]) {